% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ask_user}
\alias{ask_user}
\title{Ask the user to interactively chose which values of a tables' column to retain from a table.}
\usage{
ask_user(df, field, ask = TRUE)
}
\arguments{
\item{df}{The data.frame that will be filtered.}

\item{field}{The column name to extract the values from. It must be a character string.}
}
\value{
A filtered data.frame.
}
\description{
This function is used to let a user filter a data.frame based on the values of a column
}
\keyword{internal}
