% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests_handlers.R
\name{attempts_handler}
\alias{attempts_handler}
\title{Define errors and warnings and choose proper repeating scheme accordingly.}
\usage{
attempts_handler(input, attempts)
}
\arguments{
\item{input}{the warning/error as captured by tryCatch and withCallingHandlers}

\item{attempts}{the current number of attempts as computed by connection_handler}
}
\value{
a number defining the correct attempts value
}
\description{
Define errors and warnings and choose proper repeating scheme accordingly.
}
\keyword{internal}
