% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeMineR-package.R
\docType{package}
\name{barcodeMineR-package}
\alias{barcodeMineR}
\alias{barcodeMineR-package}
\title{barcodeMineR: Easy Mining of DNA Sequences from the NCBI and BOLD Repositories}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A comprehensive and simplified tool to query multiple taxonomic names on the National Center for Biotechnology Information (NCBI) and the Bar-code of Life Data System (BOLD) repositories, retrieving DNA sequences and associated metadata for any wanted marker. It relies on 'bold' and 'rentrez', and takes advantage of the asynchronous framework from the 'future' package to speed up the retrieval of data from the NCBI, while respecting their API requests' rate limit. The final output is a 'tibble' object, modified following the formatting requirements of the 'refdb' package, including all CDS and/or r-RNA features as distinct records with their corresponding sequence. Different quality control steps are automatically performed to deal with commonly occurring issues that derive from downloading and merging this type of data, including cleaning of mining duplicates and standardization of the different formats of BOLD and NCBI metadata (e.g. geographic coordinates, dates).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MatteoCe/barcodeMineR}
  \item \url{https://matteoce.github.io/barcodeMineR/}
  \item Report bugs at \url{https://github.com/MatteoCe/barcodeMineR/issues}
}

}
\author{
\strong{Maintainer}: Matteo Cecchetto \email{matteocecchetto@gmail.com} (\href{https://orcid.org/0000-0002-4505-4104}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Stefano Schiaparelli \email{stefano.schiaparelli@unige.it} (\href{https://orcid.org/0000-0002-0137-3605}{ORCID}) [funder]
}

}
\keyword{internal}
