% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_helpers.R
\name{bold_fetcher}
\alias{bold_fetcher}
\title{Download records from the BOLD database}
\usage{
bold_fetcher(taxon_group, bold_count, bold_tax)
}
\arguments{
\item{taxon_group}{taxonomic names}

\item{bold_tax}{data.frame output of get_bold_taxonomy}

\item{bold_counts}{data.frame output of bold_record_grouper}
}
\value{
a data.frame
}
\description{
Download records from the BOLD database
}
\keyword{internal}
