% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_helpers.R
\name{bold_record_counter}
\alias{bold_record_counter}
\title{Select the taxa with records on bold}
\usage{
bold_record_counter(bold_tax, api_rate)
}
\arguments{
\item{bold_tax}{data.frame the output of get_bold_taxonomy}

\item{api_rate}{the api_rate as set in download_bold}
}
\value{
data.frame with taxa represented by at least 1 record on bold
}
\description{
Select the taxa with records on bold
}
\keyword{internal}
