% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_helpers.R
\name{bold_record_grouper}
\alias{bold_record_grouper}
\title{Group taxonomic names by maximum cumulative value}
\usage{
bold_record_grouper(bold_count, bold_tax, rate)
}
\arguments{
\item{bold_count}{data.frame output of the bold_record_counter function}

\item{bold_tax}{data.frame output of the get_bold_taxonomy function}

\item{rate}{maximum number of records to group}
}
\value{
a list of character vectors
}
\description{
Group taxonomic names by maximum cumulative value
}
\keyword{internal}
