% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaners.R
\name{boldDuplicateCheck}
\alias{boldDuplicateCheck}
\title{Remove BOLD records if the corresponding, mined NCBI records are present}
\usage{
boldDuplicateCheck(refdb)
}
\arguments{
\item{refdb}{a refdb object}
}
\value{
a refdb object
}
\description{
Remove BOLD records if the corresponding, mined NCBI records are present
}
\keyword{internal}
