% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{buildBarcodeOre}
\alias{buildBarcodeOre}
\title{Create a refdb from the final records data.frame}
\usage{
buildBarcodeOre(records, prefix = NULL)
}
\arguments{
\item{records}{the records data.frame, as obtained by the "download" functions.}

\item{prefix}{the prefix to use in case custom IDs are chosen by the user instead of the original processID/accession numbers}
}
\value{
a refdb data.frame
}
\description{
Create a refdb from the final records data.frame
}
\keyword{internal}
