% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{buildRecord}
\alias{buildRecord}
\title{Update record information with DNA sequence data coming from the buildSequences script.}
\usage{
buildRecord(accn, records, sequences)
}
\arguments{
\item{accn}{a string in the form "accession number|marked code".}

\item{records}{the record table.}

\item{sequences}{the DNA sequences in the format "DNAStringSet".}
}
\value{
a data.frame, with information updated based on the additional information from the "built" sequences.
}
\description{
To be used after buildSequences as it calculates the final length of each marker. It only extracts the info of each chose accession number and marker combination and adds the gene_name and final length of each sequence. There is a difference depending on the origin of the records, whether they come from a search on BOLD records or NCBI records. In the former, the information obtained from the fetching of the sample data may not be unequivocally selected using the sourceID, especially for those cases where the sourceID is the same for different sequences (16S and COI CDS coming from same source sequence on the NCBI). In this case, records must be selected by sourceID and markerCode altogether as it is already available in the records table. Instead, for the NCBI the records can only be selected by sourceID, and, considering the rest of the information is unique, possible duplicate rows of the selected data.frame can be removed to only have a dataframe with a single row.
}
\keyword{internal}
