% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builders.R
\name{buildSequences}
\alias{buildSequences}
\title{Clean and process chosen sequences based on the NCBI location qualifiers.}
\usage{
buildSequences(accn, DNAString, selection_tab)
}
\arguments{
\item{accn}{a string in the form "accession number|marked code".}

\item{DNAString}{the DNA sequences in the format "DNAStringSet"}

\item{selection_tab}{the selection tab created by "extractSelectionTab"}
}
\value{
a list of "DNAStringSet" objects corresponding to the DNA sequences trimmed as indicated in the relative location qualifiers in the selection tab.
}
\description{
To be used in a vectorized loop (like lapply) for all "accession number|marker code" chosen in the selection process. This function will select the correct sequences from the total obtained in select_ncbi_genes.R or download_ncbi_genes.R and, if necessary, cut the correct region of the sequence corresponding to the chosen CDS or rRNA. It is also used inside the "select_bold_genes" function. In this case, this script will simply select and extract the "accn" previously selected from the whole sequences gathered at the beginning of the selection script and remove gaps, if present.
}
\keyword{internal}
