% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncbi_taxonomy.R
\name{clean_taxonomy}
\alias{clean_taxonomy}
\title{Collapse query names that corresponds to the same taxonomy on the ncbi.}
\usage{
clean_taxonomy(taxonomy_df)
}
\arguments{
\item{taxonomy_df}{the ncbi taxonomy table as obtained in get_ncbi_taxonomy.}
}
\value{
a taxonomy table with duplicates collapsed.
}
\description{
Sometimes it may happen that different taxonomic names correspond to only one name on the ncbi. This will collapse the rows with different query names into a single one with "name1|name2" in the field "QueryName".
}
\keyword{internal}
