% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests_handlers.R
\name{connection_handler}
\alias{connection_handler}
\title{Repeat failing functions and manage errors.}
\usage{
connection_handler(fun, attempts = 5)
}
\arguments{
\item{fun}{The same function supplied to ncbi_limit_handler.}

\item{attempts}{The number of times to repeat the function before stopping the while loop.}
}
\value{
The result of the function provided.
}
\description{
Repeat failing functions and manage errors.
}
\keyword{internal}
