% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bold.R
\name{download_bold}
\alias{download_bold}
\title{Download records from BOLD}
\usage{
download_bold(bold_tax, rate = 100, api_rate = NULL, ask = TRUE, prefix = NULL)
}
\arguments{
\item{bold_tax}{a data.frame, as returned from the get_bold_taxonomy function.}

\item{rate}{numeric, the number of taxonomic names for grouping. These will be
queried at the same time.}

\item{api_rate}{numeric, the timing rate at which to send requests. In this
function, it defaults to NULL and is always set to 1, one request per second,
due to the incompatibility of the bold functions with the future parallelizing
framework.}

\item{ask}{logical, defaults to TRUE. If disabled the selection process will
be skipped and all feature keys from each accession number will be retrieved.}

\item{prefix}{character, defaults to NULL. defaults to NULL. Character string
that will be used to create numbered custom ids for each record in ascending
order.}
}
\value{
a refdb tibble data.frame, including the DNA sequence as a field.
}
\description{
A short description...
}
