% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ncbi.R
\name{download_ncbi}
\alias{download_ncbi}
\title{Download records from the NCBI}
\usage{
download_ncbi(
  ncbi_tax = NULL,
  ncbi_ids = NULL,
  rate_xml = 200,
  rate_fasta = 100,
  default.filter = TRUE,
  filter = NULL,
  api_rate = NULL,
  ask = TRUE,
  prefix = NULL
)
}
\arguments{
\item{ncbi_tax}{a data.frame, as returned from the get_ncbi_taxonomy function.}

\item{ncbi_ids}{a character vector including ncbi accession numbers.}

\item{rate_xml}{the number of xml to be downloaded at a time. Defaults to
200, but can be lowered for bad internet connections.}

\item{rate_fasta}{the number of fasta sequences to be downloaded at a time.
Same as rate_xml, but defaults to 100. Many fasta can correspond to
mitogenomes and chromosomes, which may lead to an error if downloaded in
great numbers. Can be increased if needed.}

\item{default.filter}{logical, whether to filter the records excluding whole
genome shotgun sequences and transcribed shotgun assembly. Defaults to TRUE.}

\item{filter}{an additional query filter in the form of a/multiple string/s
to add to every searched taxid. This will allow any user to specifically
filter every search with a custom query. Multiple strings should be provided
in the form of a character vector of single query filters (see description
for details).}

\item{api_rate}{the time rate adopted for each request sent to the NCBI.
Defaults to NULL, which will let the function check whether the key parameter
is also NULL or not. In the former case, the api rate will be set to a
request every 1/3 seconds, otherwise every 1/10 seconds. Works only in a
multisession future plan. Check the "Description" for more information.}

\item{ask}{defaults to TRUE. If disabled the selection process will be
skipped and all feature keys from each accession number will be retrieved.}

\item{prefix}{defaults to NULL. Character string that will be used to create
numbered custom ids for each record in ascending order.}
}
\value{
a refdb tibble data.frame, including the DNA sequence as a field.
}
\description{
A short description...
}
