% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sequence}
\alias{example_sequence}
\title{Example DNAStringSet sequence}
\format{
## `example_sequence`
A DNAStringSet object (from Biostrings) of 658 base pairs of "width". The
item NAMES includes the sourceID and the markerCode separated by a pipe.
}
\source{
generated manually, consult data_raw directory for details
}
\usage{
example_sequence
}
\description{
An example sequence, corresponding to the data in example_record. Used as
example dataset for the function loadBarcodeOre or for the vignettes.
}
\keyword{datasets}
