% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpers.R
\name{extractRecordsTab}
\alias{extractRecordsTab}
\title{Extract records table from a list of xml source features.}
\usage{
extractRecordsTab(feature_nodes, taxonomic_table)
}
\arguments{
\item{feature_nodes}{A list of xml source features.}

\item{taxonomic_table}{The taxonomic table obtained from the get_ncbi_taxonomy function.}
}
\value{
A data.frame with source data from each xml source features, thus the metadata reported for each corresponding accession number.
}
\description{
Similar to "extractSelectionTab", this function allows to extract record information from a list of XML source features. The taxonomic table, obtained with get_ncbi_taxonomy is here inspected for taxonomic names.
}
\keyword{internal}
