% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_helpers.R
\name{extractRecordsTabBOLD}
\alias{extractRecordsTabBOLD}
\title{Extract records table from the output of bold::bold_seqspec}
\usage{
extractRecordsTabBOLD(bold_records, bold_tax)
}
\arguments{
\item{bold_records}{data.frame the output of bold::bold_seqspec}

\item{bold_tax}{data.frame the output of get_bold_taxonomy}
}
\value{
data.frame a records_tab with selected columns
}
\description{
This function is very similar to extractRecordsTab, thus allowing to extract
records information, but from the output of a bold_fetcher function, which gives
a "tsv-formatted" records table, very easily processable to create the records
tab format used here. The taxonomic table, obtained with get_bold_taxonomy is
here inspected to extract solely the BOLD taxid for the specific species processed
at each iteration.
}
\keyword{internal}
