% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpers.R
\name{extractSelectionTab}
\alias{extractSelectionTab}
\title{Extract selection table from a list of xml CDS and rRNA features.}
\usage{
extractSelectionTab(feature_nodes, accn)
}
\arguments{
\item{feature_nodes}{A list of xml CDS and rRNA features.}

\item{accn}{a character vector of accession number, those recovered using extractRecordsTab}
}
\value{
A data.frame counting the number of CDS and rRNA features.
}
\description{
This function process a list of feature XML objects in order to create the selection tab, the table that will later allow to choose which CDS and rRNA to keep from all of the accession numbers gathered. It works on a lapply function to iteratively extract data.frame-like objects for each CDS and rRNA.
}
\keyword{internal}
