% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpers.R
\name{extractTaxonomyTab}
\alias{extractTaxonomyTab}
\title{Extract taxonomic data from an NCBI xml node object.}
\usage{
extractTaxonomyTab(feature_nodes)
}
\arguments{
\item{feature_nodes}{An xml of 'XMLNodeSet' class obtained with ncbi_xml_fetcher (db = 'taxonomy').}
}
\value{
A named list
}
\description{
Works similarly to XML_extract. A named list with taxid, taxonomic rank, scientific
name, phylum, class, order, family, genus and species will be returned.
}
\keyword{internal}
