% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bold_taxonomy.R
\name{get_bold_taxonomy}
\alias{get_bold_taxonomy}
\title{Get taxonomic classification of the BOLD taxonomy database for a list of taxa}
\usage{
get_bold_taxonomy(ids, api_rate = NULL, ask = TRUE, descend = TRUE)
}
\arguments{
\item{ids}{`character` A character string of species names.}

\item{api_rate}{`integer` The API rate with which to iterate each separate
request. If left to default `NULL`, it will be set to one request every 16
seconds approximately. Use caution when overriding the default.}

\item{ask}{`logical` Should the function ask the user whether to filter the
final output for taxonomic ranks. Default `TRUE`.}

\item{descend}{`logical` Use the taxize package to retrieve lower level
taxonomies. Default `TRUE`.}
}
\value{
`data.frame` A data.frame object with the searched taxa, the taxid
  and the rank.
}
\description{
The taxonomy functions are used to define exactly which records will be
retrieved by the download functions. More precisely, only records that
include the taxid/scientificName, retrieved using the taxonomy functions,
as the lowest taxonomic identification will be downloaded.

For more details see the 'Searching taxonomy' vignette:
\code{vignette("Searching taxonomy", package = "barcodeMineR")}
}
\examples{
get_bold_taxonomy("Achelia assimilis", descend = FALSE)

}
\seealso{
[get_ncbi_taxonomy()]
}
