% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lower_tax_rank}
\alias{get_lower_tax_rank}
\title{Get lower taxonomic level from input}
\usage{
get_lower_tax_rank(tax)
}
\arguments{
\item{tax}{}
}
\value{
character vector
}
\description{
Get lower taxonomic level from input
}
\keyword{internal}
