% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncbi_taxonomy.R
\name{get_ncbi_taxonomy}
\alias{get_ncbi_taxonomy}
\title{Get taxonomic classification of the NCBI taxonomy database for a list of taxa
or taxid.}
\usage{
get_ncbi_taxonomy(ids, rate = 200, api_rate = NULL, ask = TRUE)
}
\arguments{
\item{ids}{Character string of species names or taxid from the NCBI.}

\item{rate}{the number of maximum ids to show in the esearch list object.}

\item{api_rate}{The rate with which to perform the function on each element.
Must be a number between 3 and 10 which will translate in a rate of
'1 / api_rate' seconds.}

\item{ask}{Should the function ask the user whether to filter the final
output for taxonomic ranks. Default to TRUE.}
}
\value{
A data.frame object with the searched taxonomic classification.
}
\description{
This function is the first step of the pipeline. It allows to download the
taxonomic classification of either taxonomic names or taxid from the NCBI.
A character vector of either type must be supplied to the ids parameter.
The api_rate parameter can be left NULL, as default, allowing the function
to automatically detect the best option related to the current future plan
and the NCBI http requests limit. In fact, if key is also left NULL, the
standard https requests limit will be adopted (3 requests in any 1 second
window), whereas if the NCBI API key is provided, up to 10 requests per
second will be performed. This speed up the recovery of data from the NCBI,
especially for time-consuming requests.

The future plan framework allows to send requests every 1 / api_rate seconds,
thus ensuring that no more that 'api_rate' requests will be sent. However,
due to fluctuations in the internet connection, still more than that number
of requests will arrive to the NCBI server, causing errors. The function can
handle up to 5 consecutive errors per request, but too many errors might
block the whole process. Apart from reducing the number of taxonomic names
or taxid searched, the api_rate parameter can be modified in order to slow
down the requests sent per second. It overrides the automatic selection of
the optimal parameter (either 3 or 10) and accepts one decimal degree number
between 1.0 and 10.0, so if the internet connection is particularly bad, it
can be set to 2.5/2.0, for example, in order to slow the number of requests
per second and reduce the possibility of errors.
}
\examples{
\dontrun{
get_ncbi_taxonomy("Polymastia invaginata")

get_ncbi_taxonomy("283554")

}
}
