% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_helpers.R
\name{ncbi_fasta_fetcher}
\alias{ncbi_fasta_fetcher}
\title{Fetch fasta sequences from the NCBI 'nucleotide' database}
\usage{
ncbi_fasta_fetcher(searchWebHist = NULL, ids = NULL, retstart, rate)
}
\arguments{
\item{searchWebHist}{A web_history object.}

\item{ids}{a character vector of accession/GI ids.}

\item{retstart}{The initial numeric position of the ids included in the web_history object and that will be fetched. For web_history objects with 1 ids only, 'NULL' is preferable to 0.}

\item{rate}{The final numeric position of the ids included in the web_history object and that will be fetched.}
}
\value{
A character string as fasta sequence/s.
}
\description{
Very similar to the ncbi_xml_fetcher, except that it retrieves fasta sequences.
}
\examples{
\dontrun{
 search <- ncbi_searcher("360011", db = "nucleotide")
 fetch <- ncbi_fasta_fetcher(search$web_history, retstart = 0, rate = 20)
}
}
\keyword{internal}
