% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_helpers.R
\name{ncbi_id_extract}
\alias{ncbi_id_extract}
\title{Extract ids from an esearch objects list.}
\usage{
ncbi_id_extract(search_list, db = "nucleotide", rate_xml, api_rate = NULL)
}
\arguments{
\item{search_list}{the output of the function ncbi_searcher}

\item{db}{the database against which the ids were searched.}

\item{rate_xml}{the number of xml to be downloaded at a time.}

\item{api_rate}{the time rate adopted for each request sent to the NCBI.}
}
\value{
a list with character vectors of ncbi IDs to be used by the function postAndCheck for creating web_history objects.
}
\description{
Searching with esearch the ncbi database will give web_history objects and ids. However, the number of hits corresponding to that "search" may exceed the retrieved number of ids, based on the chosen retstart and retmax (the paramete rate_xml). This will run entrez_summary in order to retrieve all ids from each esearch object.
}
\keyword{internal}
