% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_helpers.R
\name{ncbi_searcher}
\alias{ncbi_searcher}
\title{Search a taxon or taxid in the NCBI 'taxonomy' or 'nucleotide' databases}
\usage{
ncbi_searcher(
  id,
  db = "taxonomy",
  retmax = 200,
  default.filter = TRUE,
  filter = NULL
)
}
\arguments{
\item{id}{A taxon or taxid.}

\item{db}{The NCBI database that will be searched. Either 'nucleotide' or 'taxonomy'.}

\item{retmax}{the number of maximum ids to show in the esearch list object.
Defaults to 200.}

\item{default.filter}{logical, whether to filter the records excluding whole
genome shotgun sequences and transcribed shotgun assembly. Defaults to TRUE.}

\item{filter}{Additional text string/s that will be included in the query.
Works only with the 'nucleotide' option in the 'db' parameter.}
}
\value{
An esearch/list object, including a web_history object. The latter
will be used by the 'fetcher' functions to retrieve xml or fasta objects.
}
\description{
This internal function will return an esearch/list object that will be used
by the internal 'fetcher' functions to retrieve the wanted data.
It can work only on a single 'id' object (a single string character) and for
this reason is usually run inside the functions reported in "requests_handlers.R"
like connection_handler() and ncbi_limit_handler().
}
\examples{
\dontrun{
ncbi_searcher("Thouarella", db = "taxonomy")
ncbi_searcher("283554", db = "nucleotide")
}
}
\keyword{internal}
