% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_helpers.R
\name{ncbi_xml_fetcher}
\alias{ncbi_xml_fetcher}
\title{Fetch the taxonomic classification or the sequence metadata from the NCBI 'taxonomy' or 'nucleotide' databases}
\usage{
ncbi_xml_fetcher(
  searchWebHist = NULL,
  ids = NULL,
  db = "taxonomy",
  retstart,
  rate
)
}
\arguments{
\item{searchWebHist}{A web_history object.}

\item{ids}{a character vector of accession/GI ids.}

\item{db}{The NCBI database from which to fetch data. Either 'nucleotide' or 'taxonomy'.}

\item{retstart}{The initial numeric position of the ids included in the web_history object and that will be fetched. For web_history objects with 1 ids only, 'NULL' is preferable to 0.}

\item{rate}{The final numeric position of the ids included in the web_history object and that will be fetched.}
}
\value{
Both 'taxonomy' and 'nucleotide' db searches will return an xml object of 'XMLInternalDocument/XMLAbstractDocument' class. This object should be processed with the functions at 'xml_helpers.R' to extract the wanted data.
}
\description{
A web_history object is used to download xml objects from the NCBI database. The parameters retstart and rate are changed internally to download all ids from the web_history object, usually identifying 20 ids at a time.
}
\examples{
\dontrun{
  search <- ncbi_searcher("Thouarella", db = "taxonomy")
  fetch <- ncbi_xml_fetcher(search$web_history, retstart = 0, rate = 20)
}
}
\keyword{internal}
