% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaners.R
\name{ncbiDuplicateCheck}
\alias{ncbiDuplicateCheck}
\title{Remove NCBI records if the corresponding, mined BOLD records are present}
\usage{
ncbiDuplicateCheck(refdb)
}
\arguments{
\item{refdb}{a refdb object}
}
\value{
a refdb object
}
\description{
Remove NCBI records if the corresponding, mined BOLD records are present
}
\keyword{internal}
