% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plot_length}
\alias{plot_length}
\title{Create a barplot using the function geom_density_ridges from ggplot2, to compare
the final length of the sequences (field lengthGene) with the original length
of sequences (field lengthSource).}
\usage{
plot_length(
  refdb,
  limit = 1000,
  breaks = 50,
  level = NULL,
  select = NULL,
  scaling = TRUE
)
}
\arguments{
\item{refdb}{(refdb) a refdb object.}

\item{limit}{(numeric) the length in "base pairs" with which the sequences
longer than it will be grouped and reported as the last bar on the x axis.}

\item{breaks}{(numeric) the length range in "base pairs" of the x axis.}

\item{level}{(character) a character vector indicating which taxonomic level
the facet_wrap function from ggplot2 should refer to. It should be one of
"phylum", "class", "order", "family" and "genus".}

\item{select}{(character) this parameter allows to pre-filter the refdb
object based on a taxonomic name, which must be present in the records table
of the object.}

\item{scaling}{(logical) defaults to TRUE, it reports the height of the bars
according to the sequence count in the refdb object, thus, if used with the
level parameter, the heigth of bars in different wraps can be compared.}
}
\value{
a geom_density_ridges ggplot2 plot.
}
\description{
Create a barplot using the function geom_density_ridges from ggplot2, to compare
the final length of the sequences (field lengthGene) with the original length
of sequences (field lengthSource).
}
\examples{
\dontrun{
(lengthPlot <- plot_length(refdb))
}
}
