% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_helpers.R
\name{postAndCheck_OLD}
\alias{postAndCheck_OLD}
\title{Get a web_history object from a vector of accession numbers.}
\usage{
postAndCheck_OLD(names)
}
\arguments{
\item{listAccn}{a vector of accession numbers.}
}
\value{
a list with a web_history object as first element and the corresponding accessions as second element.
}
\description{
Old version. Also checks that the web_history object contains the correct number of records.
}
\keyword{internal}
