% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_accessions}
\alias{select_accessions}
\title{Filter a selection table interactively to return a vector of accession number concatenated with a gene or product name}
\usage{
select_accessions(selection_tab, ask = TRUE)
}
\arguments{
\item{selection_tab}{A selection table, as produced by the ::: function}
}
\value{
A vector of the type: "accession number|gene or product name"
}
\description{
It will take the tab with the basin info from each accession number (CDS and/or rRNA name, and coordinates of the sequence) and output a character vector with the accNum and the feature name (gene or product) separated by a pipe.
}
\keyword{internal}
