% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_id_type}
\alias{set_id_type}
\title{Is the vector a species or a taxid.}
\usage{
set_id_type(ids)
}
\arguments{
\item{ids}{The character wtring with a taxonomic name or a taxid.}
}
\value{
A condition, either TRUE or FALSE.
}
\description{
Is the vector a species or a taxid.
}
\keyword{internal}
