% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_ncbi_rate}
\alias{set_ncbi_rate}
\title{Choose the optimal http requests limit based on the presence or absence of an NCBI API key.}
\usage{
set_ncbi_rate(api_rate)
}
\arguments{
\item{api_rate}{the api rate limit, as set or not by the user}
}
\value{
A number, defining the correct api rate.
}
\description{
Choose the optimal http requests limit based on the presence or absence of an NCBI API key.
}
