% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species200}
\alias{species200}
\title{Example dataset of 200 species names}
\format{
## `species200`
A character vector of 207 species from the Ross Sea (Antarctica).
}
\source{
generated manually, consult `data_raw` directory for details
}
\usage{
species200
}
\description{
this is an example dataset, including a vector of 207 species from the Ross
Sea (Antarctica, Southern Ocean), that yield NCBI records in the numbers
between 10 and 30 (approximately), with fasta sequences never exceeding 2000
base pairs (bp) in length, and never below 100 bp (at 12 Jun 2024).
}
\keyword{datasets}
