% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{textToDNAStringSet}
\alias{textToDNAStringSet}
\title{Turn a data.frame or fasta text object of sequences into DNAStringSet}
\usage{
textToDNAStringSet(sequences)
}
\arguments{
\item{sequences}{Object containing sequences as data.frame or fasta text object}
}
\value{
DNAStringSet
}
\description{
Turn a data.frame or fasta text object of sequences into DNAStringSet
}
\keyword{internal}
