% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{web_history_parameter}
\alias{web_history_parameter}
\title{Calculate the right retstart and retmax for the NCBI fetcher functions.}
\usage{
web_history_parameter(split, hits, hits_x_iteration)
}
\arguments{
\item{split}{The starting position.}

\item{hits}{The total number of ids from the web_history object that will be used to download data from the NCBI.}

\item{hits_x_iteration}{The number of ids to fetch each execution time.}
}
\value{
A list with two values: the retstart parameter and the retmax. The last one is calculated based on the total number of ids in the web_history object and the starting position.
}
\description{
Calculate the right retstart and retmax for the NCBI fetcher functions.
}
\keyword{internal}
