% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{web_history_splitter}
\alias{web_history_splitter}
\title{Get all the retstart positions based on total number of counts and xml rate of a web_history object.}
\usage{
web_history_splitter(counts, rate)
}
\arguments{
\item{counts}{total number of records corresponding to a web_istory object}

\item{rate}{the number of record to download at eahc iteration of the ncbi_xml_fetcher function.}
}
\value{
a numeric vector with each retstart position for a particular
}
\description{
Get all the retstart positions based on total number of counts and xml rate of a web_history object.
}
\keyword{internal}
