% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpers.R
\name{XML_extract}
\alias{XML_extract}
\title{Extract GBFeature data from an NCBI xml node object.}
\usage{
XML_extract(xml, path, quals = NULL)
}
\arguments{
\item{xml}{An xml of 'XMLNodeSet' class obtained with ncbi_xml_fetcher (db = 'nucleotide').}

\item{path}{One of 'accession', 'location' or 'qualifier'. This parameter will determine what type of output will be generated.}

\item{quals}{With path set as 'qualifier' or 'location', determines which qualifiers will be extracted from the xml object. If chosen with path 'location' it must be one of 'from' or 'to'.}
}
\value{
Either a string text with the accession number (path = 'accession') or the location definition (path = 'location'), or a named list with the qualifiers extracted from GBFeature of the NCBI xml object.
}
\description{
This function will extract different data from an xml file obtained from the NCBI.
An xml corresponding to a particular feature of the entire xml obtained from the
NCBI for a specific accession number must be supplied, thus it may be of a GBFeature
corresponding to Source, CDS or rRNA. This function accepts three cases (here
chosen with the argument "path") "accession" (for the extraction of the accession
number), "location" (for the extraction of the position of the sequence in the
entire sequence) and "qualifier" (used to extract information from the source key
feature). The quals argument is used to specify different metadata the user want
to extract from the source feature key of the xml file supplied like "organism",
collected_by".
}
\keyword{internal}
