% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpers.R
\name{XML_extract_nodes}
\alias{XML_extract_nodes}
\title{Extract nodes from an NCBI nucleotide xml object.}
\usage{
XML_extract_nodes(featureKey, xml)
}
\arguments{
\item{featureKey}{What feature key to extract. Currently, "source", "CDS" and
"rRNA" are supported.}

\item{xml}{An "XMLInternalDocument", "XMLAbstractDocument" class object.}
}
\value{
An xml of 'XMLNodeSet' class.
}
\description{
Supported feature keys are "source", "CDS" and "rRNA".
}
\keyword{internal}
