% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpers.R
\name{XML_root}
\alias{XML_root}
\title{Append root of XML to new xml doc.}
\usage{
XML_root(xml)
}
\arguments{
\item{xml}{an XML object as fetched using ncbi_xml_fetcher}
}
\value{
a new xml object without the original root.
}
\description{
Append root of XML to new xml doc.
}
\keyword{internal}
