% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_helpers.R
\name{XML_Taxon_extract_nodes}
\alias{XML_Taxon_extract_nodes}
\title{Extract nodes from an NCBI taxonomy xml object.}
\usage{
XML_Taxon_extract_nodes(xml)
}
\arguments{
\item{xml}{An "XMLInternalDocument", "XMLAbstractDocument" class object.}
}
\value{
An xml of 'XMLNodeSet' class.
}
\description{
Similar function to XML_extract_nodes, but works on an NCBI taxonomy xml.
}
\keyword{internal}
