% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeMineR-package.R
\docType{package}
\name{barcodeMineR-package}
\alias{barcodeMineR}
\alias{barcodeMineR-package}
\title{barcodeMineR: Fast and reliable mining of DNA Barcodes from the NCBI and BOLD}
\description{
This package allows to query multiple taxonomic names on the NCBI and BOLD repositories and retrieve DNA barcodes and associated metadata for any wanted marker. It heavily relies on the bold and rentrez packages from the rOpenSci, and takes advantages of the parallelizing framework from the future package to speed up the retrieval of data from the NCBI, respecting their API requests limit. The final output is a tibble object modified following the formatting requirements of the refdb package, and it is cleaned from mining duplicates, differences in formats of BOLD and NCBI metadata (e.g. geographic coordinates, dates) and commonly occurring issues that derive from downloading data from those repositories. In synthesis, it provides a unified framework for mining DNA Barcodes from the main repositories, roviding clean, metadata-rich sequences in a programmatic way or interactively, depending on the wanted usage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MatteoCe/barcodeMineR}
  \item Report bugs at \url{https://github.com/MatteoCe/barcodeMineR/issues}
}

}
\keyword{internal}
