% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bold_taxonomy.R
\name{get_bold_taxonomy}
\alias{get_bold_taxonomy}
\title{Get taxonomic classification of the BOLD taxonomy database for a list of taxa.}
\usage{
get_bold_taxonomy(ids, group = 1, descend = TRUE, api_rate = 0.06, ask = TRUE)
}
\arguments{
\item{ids}{character vector of species names}

\item{group}{grouping factor for ids names. Defaults to 10.}

\item{descend}{logical. If set to TRUE (default), use the taxize package to
retrieve lower level taxonomies.}

\item{api_rate}{the rate with which to perform the function on each element.
For the BOLD api it can be higher, thus it will be defined by the number of
workers available in the chosen future framework, unless specified otherwise}

\item{ask}{should the function ask the user whether to filter the final
output for taxonomic ranks. Default to TRUE.}
}
\value{
a data.frame with taxid from the taxonomy database of the BOLD
website and the corresponding species name(s) retrieved
}
\description{
Due to the functioning of the bold package, random number generation might
set warning messages when this function is performed in multisession plan.
For this reason, the seed parameter in the future::future function called
internally is set to NULL, in order to suppress these warnings.
}
\examples{
\dontrun{
(x <- c("Achelia assimilis", "Alcyonium antarcticum"))
get_bold_taxonomy(x)
}
}
