% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ncbi_taxonomy.R
\name{get_ncbi_taxonomy}
\alias{get_ncbi_taxonomy}
\title{Get taxonomic classification of the NCBI taxonomy database for a list of taxa
or taxids}
\usage{
get_ncbi_taxonomy(ids, api_rate = NULL, ask = TRUE)
}
\arguments{
\item{ids}{`character` A character string of species names or taxid from the
NCBI.}

\item{api_rate}{`integer` The API rate with which to iterate each separate
request. Must be a number between 3 and 10 which will translate in a rate
of `1 / api_rate` seconds.}

\item{ask}{`logical` Should the function ask the user whether to filter the
final output for taxonomic ranks. Default `TRUE`.}
}
\value{
`data.frame` A data.frame object with the searched taxonomic
  classification.
}
\description{
The taxonomy functions are used to define exactly which records will be
retrieved by the download functions. More precisely, only records that
include the taxid/scientificName, retrieved using the taxonomy functions,
as the lowest taxonomic identification will be downloaded.

For more details see the 'Searching taxonomy' vignette:
\code{vignette("Searching taxonomy", package = "barcodeMineR")}
}
\examples{
get_ncbi_taxonomy("Polymastia invaginata")

get_ncbi_taxonomy("283554")

}
\seealso{
[get_bold_taxonomy()]
}
