% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBarcodeOres.R
\name{loadBarcodeOre}
\alias{loadBarcodeOre}
\title{Load sequences and corresponding records as a refdb object}
\usage{
loadBarcodeOre(pathToRecords, pathToSequences, prefix = NULL)
}
\arguments{
\item{pathToRecords}{character string with path (or file name if in the working
directory) leading to strictly "tsv file" with records information.}

\item{pathToSequences}{character string with path (or file name if in the working
directory) leading to fasta file with sequences corresponding to each record.}

\item{prefix}{defaults to NULL. Character string that will be used to create
numbered custom ids for each record in ascending order.}
}
\value{
(refdb-class) a refdb object as those created with the download_*
functions.
}
\description{
Load sequences and corresponding records as a refdb object
}
\examples{
\dontrun{
myBO <- loadBarcodeOre("path/to/table.tsv", "path/to/sequences.fasta")
}

}
