% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeBarcodeOres.R
\name{mergeBarcodeOres}
\alias{mergeBarcodeOres}
\title{Merge multiple refdb objects to create a single one.}
\usage{
mergeBarcodeOres(..., resolve.conflicts = TRUE)
}
\arguments{
\item{...}{Any number of refdb objects.}

\item{resolve.conflicts}{(default = TRUE) If set to FALSE, the script will
merge refdb objects from different sources without any quality control steps.
If left to default settings, it is used to perform a basic quality control step
required if the refdb objects were obtained from a pipeline analysis that
searched the same species in both the NCBI and BOLD databases. Records from
both databases.}
}
\value{
A refdb object, including the records and sequences from all refdb
objects provided as arguments.
}
\description{
Merge multiple refdb objects to create a single one.
}
\examples{
\dontrun{
(totBO <- mergeBarcodeOres(ncbiBO, boldBO, customBO, resolve.conflicts = TRUE))

}

}
