% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_family.R
\name{plot_primers}
\alias{plot_primers}
\title{Create a Bubble Plot showing number of sequences and average/range length for each primer}
\usage{
plot_primers(
  refdb,
  level = "phylum",
  select = NULL,
  size_range = NULL,
  measure = "average",
  tax.fct.levels = NULL,
  prim.fct.levels = NULL
)
}
\arguments{
\item{refdb}{(refdb) a refdb object.}

\item{level}{(character) a character vector indicating which taxonomic level
the x axis should refer to. It should be one of "phylum", "class", "order",
"family" and "genus".}

\item{select}{(character) this parameter allows to pre-filter the refdb
object based on a taxonomic name, which must be present in the records table
of the object.}

\item{size_range}{(integer vector) this allows to change the size of bubbles.}

\item{measure}{(character) either "range" or "average" (default). In the
first case, the range of length in the sequences of the records filtered for
each combination of primer and taxonomic name will be shown. The average
length will be shown otherwise.}

\item{tax.fct.levels}{(character) a vector of taxonomic names that allows to
re-order the x axis based on a custom order.}

\item{prim.fct.levels}{(character) a vector of primers' names that allows to
re-order the y axis based on a custom order.}
}
\value{
A Bubble plot, as a 'ggplot2' plot object, which can be saved and
modified.
}
\description{
Create a Bubble Plot showing number of sequences and average/range length for each primer
}
\examples{
\dontrun{
(BubPlot <- plot_primers(total_ore))
}
}
